requirements.txt
aiodns==3.0.0
aiofile==3.7.2
aiohttp==3.7.4.post0
anyascii==0.3.0
appdirs==1.4.4
asgiref==3.4.1
astroid==2.7.3
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.5.7
beautifulsoup4==4.9.3
black==21.7b0
cachetools==4.2.4
caio==0.9.3
certifi==2021.5.30
cffi==1.14.6
chardet==3.0.4
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.4
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2.5
django-environ==0.7.0
django-filter==2.4.0
django-modelcluster==5.1
django-taggit==1.5.1
django-treebeard==4.5.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
exponent-server-sdk==2.0.0
google-api-core==2.0.1
google-auth==2.2.1
google-cloud==0.34.0
google-cloud-secret-manager==2.7.2
googleapis-common-protos==1.53.0
grpc-google-iam-v1==0.12.3
grpcio==1.41.0
gunicorn==20.1.0
html5lib==1.1
idna==3.2
isort==5.9.3
l18n==2020.6.1
lazy-object-proxy==1.6.0
libcst==0.3.21
Markdown==3.3.4
mccabe==0.6.1
multidict==5.2.0
mypy-extensions==0.4.3
openpyxl==3.0.9
packaging==21.0
pathspec==0.9.0
Pillow==8.3.2
platformdirs==2.3.0
proto-plus==1.19.2
protobuf==3.18.0
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycares==4.0.0
pycodestyle==2.7.0
pycparser==2.20
pycryptodomex==3.10.4
PyJWT==2.1.0
pylint==2.10.2
pyparsing==2.4.7
python-decouple==3.4
pytz==2021.1
PyYAML==5.4.1
regex==2021.8.3
requests==2.26.0
rollbar==0.16.2
rsa==4.7.2
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
static3==0.7.0
stream-chat==3.13.1
tablib==3.0.0
telepath==0.2
toml==0.10.2
tomli==1.2.1
typing-extensions==3.10.0.2
typing-inspect==0.7.1
urllib3==1.26.7
wagtail==2.14.1
webencodings==0.5.1
whitenoise==5.3.0
Willow==1.4
wrapt==1.12.1
xlrd==2.0.1
XlsxWriter==1.4.5
xlwt==1.3.0
yarl==1.6.3
